+++ src/controller_transaction.h
@@ -74,6 +74,10 @@ struct controller_transaction_t{
     cvec              *ct_devices;       /* Device name -> reason (string, may be absent/NULL) */
     cvec              *ct_devices_result;/* Device name -> result (typed int32, transaction_result enum).
                                              Presence of a device here means it is/was part of the transaction. */
+    cvec              *ct_devices_synced;/* Presence of device means it successfully wrote its pulled config to
+                                            tmpdev in device_recv_config(), ie its sync-time should
+                                            be recorded as this device's running-time once
+                                            commit_pulled_devices() successfully merges tmpdev into running. */
     cxobj             *ct_devdata;       /* Generic device data, eg CS_RPC_GENERIC */
 };
 typedef struct controller_transaction_t controller_transaction;