+++ CHANGELOG.md
@@ -18,9 +18,10 @@ Expected: September 2026
 
 * Improved device error handling
   * Added a "dead socket" check to make timeout handling faster
+  * Added generation tracking of synced device datastore to avoid running containing stale device configs
 * Added a separate `connect-timeout` timer for setting up connection and syncing the config
   * This is a separate timer from `device-timeout` which is expected to be longer
-  * See: [Handling unrecoverable transaction errors (including timeouts)](https://github.com/clicon/clixon-controller/issues/247) (point 1).
+  * See: [Handling unrecoverable transaction errors (including timeouts)](https://github.com/clicon/clixon-controller/issues/247) (item 1).
   * See also: https://clixon-controller-docs.readthedocs.io/en/latest/cli.html#timeouts
 * Clarified controller transaction operations
   * Added per-device error handling of closed devices
@@ -41,7 +42,8 @@ Users may have to change how they access the system
 
 ### Corrected Bugs
 
-* Partially fixed [Pull discarded on tail-device timeout, then silently reverts the device on the next push](https://github.com/clicon/clixon-controller/issues/253) (item 2 remains)
+* Fixed: [Pull discarded on tail-device timeout, then silently reverts the device on the next push](https://github.com/clicon/clixon-controller/issues/253)
+* Fixed (partially): [Handling unrecoverable transaction errors (including timeouts)](https://github.com/clicon/clixon-controller/issues/247) (item 2 remains)
 * Fixed: double freed transaction in controller-commit sometimes caused segv
 
 ## 1.8.0